Added new ban/unban feature
Added the ability for the script to ban or unban users. In order to do this you need to execute the script with the following options: .\moe_v1.ps1 -option AddBan -SteamID <STEAMID> -timeSpan <MINUTES> -banReason <REASON>. This will update a database that we will create in the MariaDB Server for keeping track of the bans.